--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/AttributeRepository.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product;
-
-public interface AttributeRepository {
-	Attribute save(Attribute attribute);
-}
