--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/CategoryRepository.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product;
-
-public interface CategoryRepository {
-	Category save(Category category);
-}
