--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/ProductException.java
@@ -1,7 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product;
-
-public class ProductException extends IllegalArgumentException{
-	public ProductException(String message) {
-		super(message);
-	}
-}
