--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/enums/FourthCategory.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product.enums;
-
-public enum FourthCategory {
-	NORMAL;
-}
