--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/enums/PriceRange.java
@@ -1,17 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product.enums;
-
-import java.util.function.Function;
-
-public enum PriceRange {
-	ZERO(price -> price > 0), MAX(price -> price < 10000000);
-
-	private final Function<Integer, Boolean> valid;
-
-	private PriceRange(Function<Integer, Boolean> validRange) {
-		this.valid = validRange;
-	}
-
-	public boolean validRange(int price) {
-		return valid.apply(price);
-	}
-}
