--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/enums/Shipping.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product.enums;
-
-public enum Shipping {
-	NORMAL;
-}
