--- src/main/java/com/prgrms/ohouse/commerce/domain/model/product/enums/Size.java
@@ -1,5 +1,0 @@
-package com.prgrms.ohouse.commerce.domain.model.product.enums;
-
-public enum Size {
-	NORMAL;
-}
