--- src/main/java/com/prgrms/ohouse/commerce/infrastructure/repository/JpaAttributeRepository.java
@@ -1,9 +1,0 @@
-package com.prgrms.ohouse.commerce.infrastructure.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.prgrms.ohouse.commerce.domain.model.product.Attribute;
-import com.prgrms.ohouse.commerce.domain.model.product.AttributeRepository;
-
-public interface JpaAttributeRepository extends JpaRepository<Attribute, Long>, AttributeRepository {
-}
