+++ src/main/java/com/prgrms/ohouse/domain/commerce/config/QuerydslConfiguration.java
@@ -1,0 +1,18 @@
+package com.prgrms.ohouse.domain.commerce.config;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+@Configuration
+public class QuerydslConfiguration {
+	@PersistenceContext
+	private EntityManager entityManager;
+
+	@Bean
+	public JPAQueryFactory jpaQueryFactory() { return new JPAQueryFactory(entityManager);}
+}
