+++ src/main/java/com/prgrms/ohouse/domain/commerce/infrastructure/repository/JpaAttributeRepository.java
@@ -1,0 +1,9 @@
+package com.prgrms.ohouse.domain.commerce.infrastructure.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.prgrms.ohouse.domain.commerce.model.product.Attribute;
+import com.prgrms.ohouse.domain.commerce.model.product.AttributeRepository;
+
+public interface JpaAttributeRepository extends JpaRepository<Attribute, Long>, AttributeRepository {
+}
