+++ src/main/java/com/prgrms/ohouse/domain/commerce/infrastructure/repository/JpaCategoryRepository.java
@@ -1,0 +1,9 @@
+package com.prgrms.ohouse.domain.commerce.infrastructure.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.prgrms.ohouse.domain.commerce.model.product.Category;
+import com.prgrms.ohouse.domain.commerce.model.product.CategoryRepository;
+
+public interface JpaCategoryRepository extends JpaRepository<Category, Long>, CategoryRepository {
+}
