+++ src/main/java/com/prgrms/ohouse/domain/commerce/model/product/AttributeRepository.java
@@ -1,0 +1,5 @@
+package com.prgrms.ohouse.domain.commerce.model.product;
+
+public interface AttributeRepository {
+	Attribute save(Attribute attribute);
+}
