+++ src/main/java/com/prgrms/ohouse/domain/commerce/model/product/CategoryRepository.java
@@ -1,0 +1,5 @@
+package com.prgrms.ohouse.domain.commerce.model.product;
+
+public interface CategoryRepository {
+	Category save(Category category);
+}
