+++ src/main/java/com/prgrms/ohouse/domain/commerce/model/product/ProductException.java
@@ -1,0 +1,7 @@
+package com.prgrms.ohouse.domain.commerce.model.product;
+
+public class ProductException extends IllegalArgumentException{
+	public ProductException(String message) {
+		super(message);
+	}
+}
