+++ src/main/java/com/prgrms/ohouse/domain/commerce/model/product/enums/ThirdCategory.java
@@ -1,0 +1,5 @@
+package com.prgrms.ohouse.domain.commerce.model.product.enums;
+
+public enum ThirdCategory {
+	FRAME;
+}
